Fixes bug mentioned in #145 on numpy versions and moves example scenes into the package to make examples work when installing through pip.
