+++ examples/generic_holonomic.py
@@ -21,15 +21,15 @@ def run_generic_holonomic(n_steps=1000, render=False, goal=True, obstacles=True)
     ob = env.reset(pos=pos0, mount_positions=mount_positions, mount_orientations=mount_orientations)
     print(f"Initial observation : {ob}")
     if goal:
-        from examples.scene_objects.goal import dynamicGoal
+        from urdfenvs.scene_examples.goal import dynamicGoal
         env.add_goal(dynamicGoal)
 
     if obstacles:
-        from examples.scene_objects.obstacles import dynamicSphereObst2
+        from urdfenvs.scene_examples.obstacles import dynamicSphereObst2
         env.add_goal(dynamicGoal)
 
     if obstacles:
-        from examples.scene_objects.obstacles import dynamicSphereObst2
+        from urdfenvs.scene_examples.obstacles import dynamicSphereObst2
 
         env.add_obstacle(dynamicSphereObst2)
     print("Starting episode")
