+++ examples/panda_reacher.py
@@ -15,11 +15,11 @@ def run_panda(n_steps=1000, render=False, goal=True, obstacles=True):
     ob = env.reset()
     print(f"Initial observation : {ob}")
     if goal:
-        from examples.scene_objects.goal import dynamicGoal
+        from urdfenvs.scene_examples.goal import dynamicGoal
         env.add_goal(dynamicGoal)
 
     if obstacles:
-        from examples.scene_objects.obstacles import dynamicSphereObst2
+        from urdfenvs.scene_examples.obstacles import dynamicSphereObst2
 
         env.add_obstacle(dynamicSphereObst2)
     print("Starting episode")
