+++ urdfenvs/scene_examples/goal.py
