+++ urdfenvs/scene_examples/obstacle_data/duck.obj
