+++ urdfenvs/scene_examples/obstacle_data/duck.urdf
