+++ urdfenvs/scene_examples/obstacle_data/duck_vhacd.obj
