+++ urdfenvs/scene_examples/obstacle_data/sphere_goal.urdf
