+++ urdfenvs/scene_examples/obstacles.py
